fixed the header and content
