+++ src/App.js
@@ -1,42 +1,40 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 import './App.css';
 import TaskListView from './components/TaskListView';
 import AddTaskView from './components/AddTaskView';
-import Header from './components/HeaderNew';
+import Header from './components/Header';
 
 function Home() {
   return (
     <div className="container">
-      <Header />
-      <h1>🐳 Docker React App</h1>
-      <p>Welcome to your Dockerized React application!</p>
-      <p>This app is running in a Docker container with hot reload enabled.</p>
+      <h1>🦖 TaskFlow</h1>
+      <p>Welcome to TaskFlow, your Task Management Application!</p>
+      <p>This application is designed to help you manage your tasks efficiently,
+      make you more productive, and keep track of your to-do list with ease.</p>
     </div>
   );
 }
 
 function App() {
+  useEffect(() => {
+    document.title = "TaskFlow";
+  }, []);
+
   return (
     <Router>
-      <Header></Header>
-      <div className="App">
-        <nav className="navbar">
-          <Link to="/" className="nav-link">Home</Link>
-          <Link to="/addtask" className="nav-link">AddTask</Link>
-          <Link to="/tasklist" className="nav-link">ViewTask</Link>
-        </nav>
-
-        <main className="main-content">
+      <Header />
+      <div className="App d-flex flex-column min-vh-100 h-100">
+        <main className="main-content flex-grow-1 h-100">
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/addtask" element={<AddTaskView />} />
             <Route path="/tasklist" element={<TaskListView />} />
           </Routes>
         </main>
+        <Footer />
       </div>
-      <Footer />
     </Router>
   );
 }
@@ -45,7 +43,7 @@ function Footer() {
   return (
     <footer className="footer bg-dark text-light py-3 mt-auto">
       <div className="container text-center">
-        <p className="mb-0">© 2025 TaskManager | Developed by Coach CC</p>
+        <p className="mb-0">© 2025 TaskFlow | Created with React and Bootstrap</p>
       </div>
     </footer>
   );
